+++ ChangeLog.md
@@ -5,5 +5,9 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- added driver timeout configuration option
+- updated dependency `com.arangodb:arangodb-java-driver:6.16.1`
+
 ## [1.0.0] - 2021-12-11
+
 - Initial Release